Use scalaVersion from buildSettings
