+++ mill-build/src/millbuild/settings/ScalaNativeBuildSettings.scala
@@ -1,0 +1,23 @@
+package millbuild.settings
+
+import zio.{ConfigProvider, Unsafe, Runtime}
+import zio.config._
+import zio.config.magnolia.deriveConfig
+import zio.config.typesafe._
+import zio.config.yaml._
+import com.typesafe.config.ConfigFactory
+import zio.Config
+
+final case class ScalaNativeBuildSettings(
+    enable: Boolean = true,
+    version: String = ScalaNativeBuildSettings.defaultVersion
+)
+
+object ScalaNativeBuildSettings {
+  val config: Config[ScalaNativeBuildSettings] = deriveConfig[ScalaNativeBuildSettings]
+  lazy val default: ScalaNativeBuildSettings   = ScalaNativeBuildSettings()
+
+  lazy val defaultVersion = "0.4.14"
+
+  implicit lazy val rw: upickle.default.ReadWriter[ScalaNativeBuildSettings] = upickle.default.macroRW
+}
